Rename opposition to team and fix foreign key
This change renames the "opposition" table to "team" to reflect the fact
that the current team's data also needs to be in the table. That
involved a lot of changes elsewhere to refer to "team" instead of
"opposition". I also created a new "our_team" option (plus UI) to store
our team's ID, and refactored the code in TeamDataAdminAjax to use a
single function for the get functions that need to include an option ID
comparison in the JSON output.

This change also cleans up my nasty foreign key bug -- every time we ran
the table update we added redundant foreign keys. We now use consistent
names and confirm whether they exist before creating new foreign keys.