Modularise structure and use uninstall.php
This change restructures the existing code to use a more modular
approach. I have split the core functionality that was in the TeamData
class into TeamDataBase and TeamDataInstaller, as well as making use of
the PHP spl_autoload_register() functionality instead of using manual
includes. Finally, I have created an uninstall.php file for the plugin
instead of an uninstall hook. The change also makes sure that all
options are removed in addition to the tables.