Rename "role" as "list"
The "role" table, and all the nomenclature surrounding it was really
inaccurate. The name should always have included some sort of email
reference, as the original goal was to support multiple simple mailing
lists.

This change renames the table and all the text and functions that refer
to roles.

Version number bumped to v 0.31