Add active flag to member table
The active flag will be used to indicate whether a user should continue
to receive email. In some cases (including my primary use case) we want
to hang on to the contact information, so we don't want to simply delete
the user. Actual manipulation of the new value will occur with changes
to the Members admin UI.

Version bumped to 0.32.