Implement initial member management
Actually implement the interfaces and UI for managing members. This
includes a change to generalize the classes and CSS for tables in the
admin pages. It also implements the code to allow members to be
automatically enrolled for specific mailing lists, as well as
controlling whether mail can be sent to all active members.

Includes a version bump to 0.33 for new columns in the list table.