Make email transmissions work
This change actually gets outbound email transmission working and
configurable. It seems to be working with SMTP and STARTTLS, but YMMV
with other PHPMailer options. There may still be more performance work
to be done, as well as testing larger numbers of recipients, but that
will need to follow separately.

The gist of the changes are as follows:
- Split off admin options to a separate section of the admin UI (with
the email enabled flag controlling whether the options are displayed at
all), as well as bulking up on the options themselves.
- Build out the UI for sending emails to specific lists
- Allow a specific From email and display to be specified for specific
email lists
- Fix option AJAX code to correctly strip slashes from incoming option
data
- Fix saving of SMTP credentials and ensure that we don't display the
password any longer than necessary
- Build out the back end for sending email